# core-banking/frontned/web-saas/saas-api-docs.json
{"openapi":"3.1.0","info":{"title":"SAAS-API","description":"SAAS-API 명세서","version":"V1"},"servers":[{"url":"http://localhost:8091","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/api/v1/auth/refresh":{"post":{"tags":["auth-controller"],"operationId":"refresh","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/com.revy.example.jwt.payload.RefreshTokenRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/auth/logout":{"post":{"tags":["auth-controller"],"operationId":"logout","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/com.revy.example.jwt.payload.LogoutRequest"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/com.revy.example.jwt.payload.LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/accounts":{"get":{"tags":["account-controller"],"operationId":"search","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/org.springdoc.core.converters.models.Pageable"}},{"name":"request","in":"query","required":true,"schema":{"$ref":"#/components/schemas/SaasAccountPayload.SearchRequest"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}},"post":{"tags":["account-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaasAccountPayload.CreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/accounts/{accountId}/withdraw":{"post":{"tags":["account-controller"],"operationId":"withdraw","parameters":[{"name":"accountId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaasAccountPayload.WithdrawRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/accounts/{accountId}/trades/sell":{"post":{"tags":["trade-controller"],"operationId":"sell","parameters":[{"name":"accountId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaasTradePayload.SellRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/accounts/{accountId}/trades/dividend":{"post":{"tags":["trade-controller"],"operationId":"dividend","parameters":[{"name":"accountId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaasTradePayload.DividendRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/accounts/{accountId}/trades/buy":{"post":{"tags":["trade-controller"],"operationId":"buy","parameters":[{"name":"accountId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaasTradePayload.BuyRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/accounts/{accountId}/deposit":{"post":{"tags":["account-controller"],"operationId":"deposit","parameters":[{"name":"accountId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaasAccountPayload.DepositRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/accounts/{accountId}":{"get":{"tags":["account-controller"],"operationId":"get","parameters":[{"name":"accountId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}},"delete":{"tags":["account-controller"],"operationId":"close","parameters":[{"name":"accountId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}},"patch":{"tags":["account-controller"],"operationId":"update","parameters":[{"name":"accountId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaasAccountPayload.UpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/stocks":{"get":{"tags":["stock-controller"],"operationId":"search_1","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/org.springdoc.core.converters.models.Pageable"}},{"name":"request","in":"query","required":true,"schema":{"$ref":"#/components/schemas/SaasStockPayload.SearchRequest"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/stocks/{stockId}":{"get":{"tags":["stock-controller"],"operationId":"get_1","parameters":[{"name":"stockId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/auth/me":{"get":{"tags":["auth-controller"],"operationId":"me","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/accounts/{accountId}/transactions":{"get":{"tags":["account-controller"],"operationId":"searchTransactions","parameters":[{"name":"accountId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/org.springdoc.core.converters.models.Pageable"}},{"name":"request","in":"query","required":true,"schema":{"$ref":"#/components/schemas/SaasAccountTxPayload.SearchRequest"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/accounts/{accountId}/transactions/{transactionId}":{"get":{"tags":["account-controller"],"operationId":"getTransaction","parameters":[{"name":"accountId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"transactionId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/accounts/{accountId}/positions":{"get":{"tags":["account-controller"],"operationId":"getPositions","parameters":[{"name":"accountId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/accounts/{accountId}/positions/{stockId}":{"get":{"tags":["account-controller"],"operationId":"getPosition","parameters":[{"name":"accountId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"stockId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/accounts/{accountId}/portfolio":{"get":{"tags":["portfolio-controller"],"operationId":"get_2","parameters":[{"name":"accountId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}}},"components":{"schemas":{"com.revy.example.jwt.payload.RefreshTokenRequest":{"type":"object","properties":{"refreshToken":{"type":"string","minLength":1}},"required":["refreshToken"]},"ApiResponse":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/com.revy.example.saas.auth.payload.UserAuthResponse"},"message":{"type":"string"},"timestamp":{"type":"integer","format":"int64"}}},"com.revy.example.saas.auth.payload.UserAuthResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"tokenType":{"type":"string"}}},"com.revy.example.jwt.payload.LogoutRequest":{"type":"object","properties":{"refreshToken":{"type":"string"}}},"com.revy.example.jwt.payload.LoginRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","minLength":1}},"required":["email","password"]},"SaasAccountPayload.CreateRequest":{"type":"object","properties":{"accountName":{"type":"string","minLength":1},"accountType":{"type":"string","enum":["REAL","VIRTUAL"]},"currency":{"type":"string","minLength":1}},"required":["accountName","accountType","currency"]},"SaasAccountPayload.WithdrawRequest":{"type":"object","properties":{"amount":{"type":"number","minimum":0.0},"referenceId":{"type":"string","minLength":1}},"required":["amount","referenceId"]},"SaasTradePayload.SellRequest":{"type":"object","properties":{"stockId":{"type":"integer","format":"int64"},"quantity":{"type":"number","minimum":0.0},"price":{"type":"number","minimum":0.0},"fee":{"type":"number","minimum":0.0},"tax":{"type":"number","minimum":0.0},"referenceId":{"type":"string","minLength":1},"tradedAt":{"type":"string","format":"date-time"}},"required":["fee","price","quantity","referenceId","stockId","tax","tradedAt"]},"SaasTradePayload.DividendRequest":{"type":"object","properties":{"stockId":{"type":"integer","format":"int64"},"grossAmount":{"type":"number","minimum":0.0},"tax":{"type":"number","minimum":0.0},"referenceId":{"type":"string","minLength":1},"tradedAt":{"type":"string","format":"date-time"}},"required":["grossAmount","referenceId","stockId","tax","tradedAt"]},"SaasTradePayload.BuyRequest":{"type":"object","properties":{"stockId":{"type":"integer","format":"int64"},"quantity":{"type":"number","minimum":0.0},"price":{"type":"number","minimum":0.0},"fee":{"type":"number","minimum":0.0},"tax":{"type":"number","minimum":0.0},"referenceId":{"type":"string","minLength":1},"tradedAt":{"type":"string","format":"date-time"}},"required":["fee","price","quantity","referenceId","stockId","tax","tradedAt"]},"SaasAccountPayload.DepositRequest":{"type":"object","properties":{"amount":{"type":"number","minimum":0.0},"referenceId":{"type":"string","minLength":1}},"required":["amount","referenceId"]},"SaasAccountPayload.UpdateRequest":{"type":"object","properties":{"accountName":{"type":"string","minLength":1}},"required":["accountName"]},"org.springdoc.core.converters.models.Pageable":{"type":"object","properties":{"page":{"type":"integer","format":"int32","minimum":0},"size":{"type":"integer","format":"int32","minimum":1},"sort":{"type":"array","items":{"type":"string"}}}},"SaasStockPayload.SearchRequest":{"type":"object","properties":{"ticker":{"type":"string"},"name":{"type":"string"},"exchange":{"type":"string"},"sector":{"type":"string"},"currency":{"type":"string"}}},"SaasAccountPayload.SearchRequest":{"type":"object","properties":{"accountNumber":{"type":"string"},"accountType":{"type":"string","enum":["REAL","VIRTUAL"]},"status":{"type":"string","enum":["ACTIVE","SUSPENDED","CLOSED"]},"currency":{"type":"string"}}},"SaasAccountTxPayload.SearchRequest":{"type":"object","properties":{"stockId":{"type":"integer","format":"int64"},"txType":{"type":"string","enum":["DEPOSIT","WITHDRAWAL","BUY","SELL","DIVIDEND","FEE","TAX"]},"status":{"type":"string","enum":["PENDING","COMPLETED","CANCELLED","FAILED"]},"referenceId":{"type":"string"},"tradedAtFrom":{"type":"string","format":"date-time"},"tradedAtTo":{"type":"string","format":"date-time"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}